requirements.txt
asn1crypto==1.4.0
attrs==21.2.0
awswrangler==2.9.0
beautifulsoup4==4.9.3
boto3==1.18.0
botocore==1.21.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.2
click==8.0.1
et-xmlfile==1.1.0
idna==2.10
importlib-metadata==4.6.1
iniconfig==1.1.1
jmespath==0.10.0
lxml==4.6.3
numpy==1.20.3
openpyxl==3.0.7
packaging==21.0
pandas==1.2.5
pep517==0.10.0
pg8000==1.19.5
pip-tools==6.2.0
pluggy==0.13.1
py==1.10.0
py-openaq==1.1.0
py4j==0.10.9
pyarrow==4.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
pyspark==3.1.2
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
redshift-connector==2.0.882
requests==2.25.1
s3transfer==0.5.0
scramp==1.4.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
